Show placeholder while updating the cart
