+++ client/my-sites/checkout/upsell-nudge/purchase-modal/constants.js
@@ -1,0 +1,1 @@
+export const BEFORE_SUBMIT = 'before-submit';
